Add support for storing test step content into an external file as well as into the XML project file.
Feature can be turned on or off with a setting.  Another setting makes
step content appear in the XML project file also; this is for backward
compatibility of project file with soapui version not having the feature.
A third setting auto-converts projects to use this feature (conversion is
done when a project is opened).

The file name can be automatically computed, composed from pre-defined
parts of manually provided.  Auto-conversion selects the AUTO mode for
computing the file name. In AUTO mode, the file name is deduced from the
directory where the project file is and the path to the step inside the
project.  A settings panel on every step is available to control these
aspects.